+++ src/common/utils/parseBoolean.ts
@@ -1,4 +1,12 @@
-export default function parseBoolean(strBool: never): boolean {
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+/**
+ * Parses a non boolean value into a boolean value
+ * @example
+ *  parseBoolean('true') => true
+ *  parseBoolean(0) => false
+ */
+export default function parseBoolean(strBool: any): boolean {
   if (!['int', 'string'].includes(typeof strBool)) {
     return false;
   }