+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.4.0-rc
 
-- Added a minimize to system tray feature
+- Added a configurable minimize to system tray feature
+- Deprecated the updateFrontEndTheme method in favor of the generic updateSettings method
 
 # 0.3.0